Remove +1 offset
Previously, this module offset the syncing by 1 so that you wouldn't get a situation where it continuously calls `wait` with smaller and smaller numbers until it passes the threshold where it can then call `wait` with higher numbers again. Now that we use `math.ceil` for the calculations, this is no longer a problem.